requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
httpx
PyJWT[crypto]
cryptography
jinja2
python-multipart
pydantic-settings
